% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pca-names.R
\name{get_pca_names}
\alias{get_pca_names}
\title{Get PCA Names}
\usage{
get_pca_names(x, type = c("rotation", "projection"), dim = 1L, value)
}
\arguments{
\item{x}{A \code{pca} class object. See \code{\link[=pca]{pca()}}.}

\item{type}{\code{character(1)}. Matched. Either \code{"rotation"} or \code{"projection"}.}

\item{dim}{\code{integer(2)}. Which PCA dimension to select.}

\item{value}{\code{numeric(1)}. The value corresponding to the
dimension to interrogate for names.}
}
\value{
A character string corresponding to the feature names
that exceed \code{value} in the dimension \code{dim}.
}
\description{
Determine the names of the features (rotation) or samples
(projection) above/below a specified cutoff (value) in the
principal component dimension specified by \code{dim}.
}
\examples{
pca <- pca(pcapkg:::log10_ft(simdata))
get_pca_names(pca, "r", 1, 0.1)      # feature names

get_pca_names(pca, "p", value = 0.1) # sample names
}
\seealso{
\code{\link[=pca]{pca()}}
}
\author{
Stu Field
}
