% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTopPCArotation.R
\name{getTopPCArotation}
\alias{getTopPCArotation}
\title{Get Top Features by PCA Dimension}
\usage{
getTopPCArotation(x, dims = 1:5, n = 30L)
}
\arguments{
\item{x}{A \code{pca} class object.}

\item{dims}{Integer. Vector of dimensions to interrogate.}

\item{n}{Integer. The limit on how many features
to report (i.e. the rows of the resulting data frame).}
}
\value{
A data frame of the top features ordered by the principal
components given in the selected dimension.
}
\description{
Interrogates the rotation in each dimension defined by \code{dims}
and orders the features by the value of the principal
components in that dimension. Each dimension is
represented by a column of the resulting data frame.
}
\examples{
pca <- pca(log10(sim_test_data))
getTopPCArotation(pca, n = 10)
}
\author{
Michael R. Mehan
}
