% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{package}
\name{objects}
\alias{objects}
\alias{smvs}
\alias{sh_list}
\title{Package Objects}
\format{
A list containing 4 vectors:
\itemize{
\item{SerumCellAbuse}
\item{PlasmaCellAbuse}
\item{PlasmaPlatelet}
\item{Complement}
}

A list containing 4 vectors:
\itemize{
\item{Complement}
\item{Platelet}
\item{plasmaLysis}
\item{serumLysis}
}
}
\source{
\code{smvs}: Mike Mehan

\code{sh_list}: Mike Mehan
}
\usage{
smvs

sh_list
}
\description{
This package contains the general functions necessary for the unsupervised exploratory analysis of SOMAmer data using Principal Component Analysis (PCA). Primary tools related to calculation of principal components and the visualization of loadings and projections of these components.

The sample mapping vectors related to sample handling.

The "official" \code{SeqIds} of the V3 sample handling lists.
}
\examples{
# ------ SomaPCA -------
lapply(smvs, head)

lapply(sh_list, head)
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{datasets}
