% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca.R
\name{pca}
\alias{pca}
\alias{print.pca}
\alias{plot.pca}
\alias{plot_scree}
\title{Principal Component Analysis}
\usage{
pca(data, features = NULL, center = TRUE, scale = FALSE)

\method{print}{pca}(x, ...)

\method{plot}{pca}(
  x,
  type = c("projection", "rotation"),
  dims = 1:2L,
  color,
  identify = FALSE,
  id.labels,
  ...
)

plot_scree(
  x,
  n = min(15L, length(attributes(x)$sdev)),
  type = c("barplot", "lines")
)
}
\arguments{
\item{data}{A \code{data.frame} or \code{tibble} class object.}

\item{features}{Which colunns are the features.}

\item{center}{Logical. Should the features/variables be zero centered
prior to decomposition? In general this should be performed so that
variables are in the same space, thus defaults to \code{TRUE}.}

\item{scale}{Logical. Should the features be scaled to have
unit variance prior to decomposition? If \code{TRUE}, this corresponds
to PCA in correlation space as opposed to covariance space.}

\item{x}{A \code{pca} class object.}

\item{...}{Additional arguments passed to \code{\link[=geom_point]{geom_point()}}.}

\item{type}{Either \code{projection} (default) or \code{rotation}. If called from
\code{\link[=plot_scree]{plot_scree()}}, either \code{barplot} (default) or \code{lines}.}

\item{dims}{\code{integer(2)}. Which dimensions to plot.}

\item{color}{An unquoted string indicating the variable in
\code{x$projection} to color points by. Required for \code{projection} plots.}

\item{identify}{Logical. Identify points? For \code{rotation} space, the top
\code{10} target names in each dimension are identified.}

\item{id.labels}{An unquoted string indicating the variable in
\code{x$projection} containing point labels. Ignored for \code{rotation} plots.}

\item{n}{\code{integer(1)}. The number of components to plot.}
}
\value{
An object of class \code{pca}, similar to \code{\link[=prcomp]{prcomp()}}.
}
\description{
Perform principal component analysis (PCA) on a proteomic data matrix,
via the decomposition of the variance-covariance matrix (\code{SVD}).
Some modification of the standard \code{\link[=prcomp]{prcomp()}} is performed.
Data transformations (e.g. centering and/or scaling) can be performed
via its arguments.

The \code{\link[=plot_scree]{plot_scree()}} function generates a \code{"scree plot"}
of a \code{pca} class object.
}
\section{Methods (by generic)}{
\itemize{
\item \code{print(pca)}: The S3 print method for objects of class \code{pca}.

\item \code{plot(pca)}: The S3 plot method for objects of class \code{pca}.

}}
\examples{
pca <- pca(pcapkg:::log_rfu(simdata))

# print method
pca

# S3 plot methods
plot(pca, "r")                         # default rotation
plot(pca, "r", identify = TRUE)        # label targets

plot(pca, "p", color = class_response) # default samples projection
plot(pca, "p", color = gender)         # use "gender" for colors
plot(pca, "p", color = class_response, id.labels = id) # label points
plot(pca, "p", color = class_response, alpha = 0.5)    # adjust point alpha
# Scree plots
plot_scree(pca)               # barplot
plot_scree(pca, type = "l")   # lines
plot_scree(pca, type = "l")
}
\seealso{
\code{\link[=prcomp2]{prcomp2()}}
}
\author{
Stu Field
}
