% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcapkg-package.R
\docType{package}
\name{pcapkg-package}
\alias{pcapkg}
\alias{pcapkg-package}
\title{pcapkg: Perform PCA Analysis of Data}
\description{
This package contains the general functions necessary for the unsupervised exploratory analysis of proteomic data using Principal Component Analysis (PCA). Primary tools related to calculation of principal components and the visualization of loadings and projections of these components.
}
\author{
\strong{Maintainer}: Stu Field \email{stu.g.field@gmail.com} (\href{https://orcid.org/0000-0002-1024-5859}{ORCID}) [copyright holder]

}
\keyword{internal}
\keyword{package}
