% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rotation.R
\name{plot_rotation}
\alias{plot_rotation}
\title{Plot PCA Rotation}
\usage{
plot_rotation(
  data.prcomp,
  dims = 1:2L,
  classes = NULL,
  scores = NULL,
  col = NULL,
  set1 = NULL,
  set2 = NULL,
  set3 = NULL,
  set4 = NULL,
  set5 = NULL,
  lab_cex = 3,
  pt_cex = 2.5,
  auto_ident = TRUE,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object
returned by \code{\link[=prcomp2]{prcomp2()}}.}

\item{dims}{\code{integer(2)}. Which dimensions to plot.}

\item{classes}{Optional. A vector indicating the classes of features used for
coloring the points. Must be the same length as the number of features.}

\item{scores}{Optional. Statistical scores to pass through for the
coloring of the points during plotting. If a training data
set is passed and \code{scores = NULL},
then KS-distances (scores) will be calculated under the
hood and used to determine point color. Can be either a
single value (e.g. "red") or a vector of color values the
same length as the number of observations.
This overrides the point color assigned by \code{classes} above.}

\item{col}{\code{character(1)}. The color of the points. Can be either
a single value (e.g. "red") or a vector of color values the
same length as the number of observations.
This parameter overrides the point colors determined by
both the \code{classes} and \code{scores} parameters above.}

\item{set1}{Optional. A vector of feature ids to mark on the plot. Typically
a vector of analytes of clinical relevance, e.g. sample handling.
Marked with a hollow diamond (see \code{\link[=pch]{pch()}}).}

\item{set2}{Optional. An additional vector of feature ids to mark
on the plot, separately from those in \code{set1}. Marked with a hollow
triangle (see \code{\link[=pch]{pch()}}).}

\item{set3}{Optional. An additional vector of feature ids to mark
on the plot, separately from those in \code{set1}. Marked with a hollow
square (see \code{\link[=pch]{pch()}}).}

\item{set4}{Optional. An additional vector of feature ids to mark
on the plot, separately from those in \code{set1}. Marked with a hollow
circle (see \code{\link[=pch]{pch()}}).}

\item{set5}{Optional. An additional vector of feature ids to mark
on the plot, separately from those in \code{set1}. Marked with a hollow
upside-down triangle (see \code{\link[=pch]{pch()}}).}

\item{lab_cex}{\code{numeric(1)}. Font size of the point labels generated by \code{auto_ident}.
\code{auto_ident} must be TRUE.}

\item{pt_cex}{\code{numeric(1)}. Character expansion for the points.}

\item{auto_ident}{\code{logical(1)}. Should the top 10 features, in both dimensions,
be identified and labeled on the plot?}

\item{...}{Additional arguments passed to \code{\link[=plot_pca_dims]{plot_pca_dims()}}.}
}
\description{
Plot the features from a PCA rotation in a 2-dimensional scatter plot.
\strong{Note:} The precedence for the ordering of the coloring is as follows:
\enumerate{
\item{\code{col}}
\item{\code{scores}}
}
}
\examples{
pca <- center_scale(pcapkg:::log10_ft(simdata), center = TRUE, scale = FALSE) |>
  feature_matrix() |>
  prcomp2()
plot_rotation(pca, col = "green")

class <- withr::with_seed(123, sample(simdata$class_response, 40L))
plot_rotation(pca, classes = class)

y <- withr::with_seed(123, sample(pcapkg:::get_analytes(simdata), 10L))
plot_rotation(pca, set1 = y)
}
\seealso{
Other PCA plots: 
\code{\link{plot_pca_dims}()},
\code{\link{plot_projection}()},
\code{\link{screeplot_auc}()}
}
\author{
Michael Mehan
}
\concept{PCA plots}
