% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-sample-handling.R
\name{plot_sample_handling}
\alias{plot_sample_handling}
\title{Plot Sample Handling (PCA)}
\usage{
plot_sample_handling(
  data.prcomp,
  extra_feat = NULL,
  samples,
  file = NULL,
  matrix_type,
  dims = 1:2L,
  extra_feat_col = "firebrick3",
  legend_pos = c(1L, 1L)
)
}
\arguments{
\item{data.prcomp}{A \code{prcomp} object, the result of
\code{supervised_peel()} or \code{\link[=prcomp2]{prcomp2()}}. If the result of a
call to \code{supervised_peel()}, any 1 of the 4 components may be
plotted, e.g. \code{orig}, \code{weighted}, \code{unweighted}, or \code{peeled}.}

\item{extra_feat}{\code{character(n)}. vector of additional features
to highlight, beyond the sample handling SMVs.}

\item{samples}{Character. A string corresponding to the samples to be used
to color the points in the projection plot.}

\item{file}{File name if plot is to be saved to an external file.}

\item{matrix_type}{\code{character(1)}. Either "plasma" or "serum".}

\item{dims}{\code{integer(2)}. Which dimensions of the PCA to plot. Must be
the same for the projection and rotation.}

\item{extra_feat_col}{\code{character(1)}. The color to be used for
the additional features specified in \code{extra_feat} argument.}

\item{legend_pos}{Coordinates of the two legends, see Section below.}
}
\value{
A \code{ggplot2} plot.
}
\description{
Quick wrapper for plotting PCA sample handling, both the sample projection
and the rotation (loadings) of the protein space.
}
\section{Legend Positions}{

\tabular{rl}{
\code{c(1, 1)} \tab top right \cr
\code{c(1, 0)} \tab bottom right \cr
\code{c(0, 1)} \tab top left \cr
\code{c(0, 0)} \tab bottom left\cr
}
}

\examples{
data <- pcapkg:::log_rfu(simdata[c(1:10L, 91:100L), ])
# Create fake matching SH features
new_seqs <- c(
  "seq.4124.24",   # cell lysis
  "seq.2811.27",   # platelet activation
  "seq.2242.51"    # complement activation
)
names(data)[31:33L] <- new_seqs   # assign into 'noise' features
pca <- center_scale(data, center = TRUE, scale = FALSE) |>
  feature_matrix() |>
  prcomp2()

plot_sample_handling(pca, samples = data$class_response, matrix_type = "p")
plot_sample_handling(pca, samples = data$class_response, matrix_type = "p",
                     extra_feat = "seq.2802.68")

# Save to file
f_out <- tempfile("plot-", fileext = ".pdf")
plot_sample_handling(pca, samples = data$class_response, matrix_type = "p",
                   legend_pos = c(0, 0), file = f_out)
}
\seealso{
Other PCA plots: 
\code{\link{plot_pca_dims}()},
\code{\link{plot_projection}()},
\code{\link{plot_rotation}()},
\code{\link{screeplot_auc}()}
}
\author{
Stu Field
}
\concept{PCA plots}
