% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPCAdims.R
\name{plotPCAdims}
\alias{plotPCAdims}
\title{Plot A PCA Dimension}
\usage{
plotPCAdims(
  data.prcomp,
  value = c(NA, "rotation", "x"),
  dims,
  classes,
  main = sprintf("PCA Plot (\%s)", ifelse(value == "x", "projection", value)),
  scores = NULL,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  do.legend = deprecated(),
  legend.pos = "right",
  xlim = NULL,
  ylim = NULL,
  colors = deprecated(),
  pt.cex = 2.5,
  pt.pch = 19,
  pt.bg = deprecated(),
  add.ellipse = FALSE,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object returned by
\code{\link[=prcomp2]{prcomp2()}}.}

\item{value}{Character. The type of plot to be generated. For projection
enter "x", for rotation enter "rotation".}

\item{dims}{Integer. Length 2. Which dimensions to plot.}

\item{classes}{Optional. A vector indicating the classes of
observations (either features or samples) used for coloring the
points, depending on whether a rotation or a projection is to be
plotted. Must be the same length as the plot type, i.e. number
of samples for projection and features for rotation.}

\item{main}{Character. A string containing the title for the plot.}

\item{scores}{Optional. KS scores to pass through for the coloring
of the points during plotting. If a training data set is passed and
\code{scores = NULL}, then KS-distances (scores) will be calculated under the
hood and used to determine point color. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This overrides the point color assigned by \code{classes} above.}

\item{col}{Character. The color of the points. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This parameter overrides the point colors determined by
both the \code{classes} and \code{scores} parameters above.}

\item{xlab}{Character. A string indicating the label of the x axis.}

\item{ylab}{Character. A string indicating the label of the y axis.}

\item{do.legend}{Deprecated. Should a legend be plotted? Particularly
for a projection.}

\item{legend.pos}{Character. Position of the legend, if plotted.
Options include "left", "right", "bottom", or "top", or "none"
if no legend is desired. Ignored if \code{value = "r"} (rotation) and
classes are not specified.}

\item{xlim}{Numeric. Optional limits for the x-axis in the format \code{c(0, 0)}.}

\item{ylim}{Numeric. Optional limits for the y-axis in the format \code{c(0, 0)}.}

\item{colors}{Deprecated. A string of colors for the points.
If \code{NULL}, reverts to a default palette string.}

\item{pt.cex}{Numeric. Character expansion for the points.}

\item{pt.pch}{Numeric. Shape of the points. Accepted values are 0-25.}

\item{pt.bg}{Deprecated. Background color for the points.}

\item{add.ellipse}{Logical. Should an ellipse be added to the
rotation? Ignored if \code{value = "x"} (projection).}

\item{...}{Additional arguments passed to \code{\link[=geom_point]{geom_point()}}.}
}
\value{
A points plot of either rotation or projection space.
}
\description{
Plots either the projection or the loadings (rotation) of
\code{"prcomp"} object. This is a base function that is
wrapped into driver parent functions in the plotting of
principal components, most notably \code{\link[=plotRotation]{plotRotation()}}
and \code{\link[=plotProjection]{plotProjection()}}.
}
\examples{
# Prepare data
pca <- center_scale(log10(sim_test_data), center = TRUE, scale = FALSE) |>
  strip_meta() |>
  prcomp2()
scores <- withr::with_seed(101, rnorm(pcapkg:::getAnalytes(sim_test_data, n = TRUE)))

# Rotation space, with scores to define point colors
plotPCAdims(pca, value = "r", dims = 1:2L, classes = NULL,
            scores = rnorm(pcapkg:::getAnalytes(sim_test_data, n = TRUE)))

# Projection space, with classes to define point colors
plotPCAdims(pca, value = "x", dims = 1:2L,
            classes = sim_test_data$class_response)
}
\seealso{
Other PCA plots: 
\code{\link{plotProjection}()},
\code{\link{plotRotation}()},
\code{\link{plotSampleHandling}()},
\code{\link{screeplotAUC}()}
}
\author{
Michael Mehan, Amanda Hiser
}
\concept{PCA plots}
