% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotPointsEllipse.R
\name{plotPointsEllipse}
\alias{plotPointsEllipse}
\title{Plot Points Ellipse (deprecated)}
\usage{
plotPointsEllipse(x, add = TRUE, scale = 1, ...)
}
\arguments{
\item{x}{A matrix (n x 2) decomposition of either
the projection ("x") or rotation ("rotation") from
a \code{prcomp} object.}

\item{add}{Logical. Should the ellipse plot be added
to an existing plot?}

\item{scale}{Numeric. Scale shift for the mean.}

\item{...}{Additional arguments.}
}
\description{
Create a points ellipse plot.
}
\examples{
\dontrun{
  pca <- center_scale(log10(sim_adat), center = TRUE, scale = FALSE) |>
    strip_meta() |> prcomp2()
  plotPointsEllipse(pca$rotation[, 1:2L], add = FALSE, lwd = 2, col = "dodgerblue")
}
}
