% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProjection.R
\name{plotProjection}
\alias{plotProjection}
\title{Plot PCA Projection}
\usage{
plotProjection(
  data.prcomp,
  dims = 1:2L,
  classes = NULL,
  scores = NULL,
  col = NULL,
  samples = NULL,
  pt.cex = 2.5,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object returned by
\code{\link[=prcomp2]{prcomp2()}}.}

\item{dims}{Integer. Length 2. Which dimensions to plot.}

\item{classes}{Optional. A vector indicating the classes of samples used for
coloring the points. Must be the same length as the number of samples.}

\item{scores}{Optional. KS scores to pass through for the coloring
of the points during plotting. If a training data set is passed and
\code{scores = NULL}, then KS-distances (scores) will be calculated under the
hood and used to determine point color. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This overrides the point color assigned by \code{classes} above.}

\item{col}{Character. The color of the points. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This parameter overrides the point colors determined by
both the \code{classes} and \code{scores} parameters above.}

\item{samples}{Optional. A vector of sample IDs to mark on the projection plot.
Specified samples are marked with a hollow diamond (see \code{\link[=pch]{pch()}}).}

\item{pt.cex}{Numeric. Character expansion for the points.}

\item{...}{Additional arguments passed to \code{\link[=plotPCAdims]{plotPCAdims()}}.}
}
\description{
Plot the samples from a PCA projection in a 2-dimensional
scatter plot.
}
\examples{
pca <- centerScaleData(log10(sim_test_data), center = TRUE, scale = FALSE) |>
  stripMeta() |>
  prcomp2()

# Define color of points
plotProjection(pca, col = "green")

# Use classes to define point colors
plotProjection(pca, classes = sim_test_data$class_response)

# Mark specific samples
plotProjection(pca, samples = rownames(sim_test_data)[1:5])
}
\seealso{
Other PCA plots: 
\code{\link{plotPCAdims}()},
\code{\link{plotRotation}()},
\code{\link{plotSampleHandling}()},
\code{\link{screeplotAUC}()}
}
\author{
Michael Mehan, Amanda Hiser
}
\concept{PCA plots}
