% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSampleHandling.R
\name{plotSampleHandling}
\alias{plotSampleHandling}
\title{Plot Sample Handling (PCA)}
\usage{
plotSampleHandling(
  data.prcomp,
  add.apts = NULL,
  samples,
  filename = NULL,
  scale = 1,
  matrix.type,
  dims = 1:2L,
  add.apt.col = "firebrick3",
  legend.pos = c(1, 1)
)
}
\arguments{
\item{data.prcomp}{A \code{prcomp} object, the result of
\code{supervised_peel()} or \code{\link[=prcomp2]{prcomp2()}}. If the result of a
call to \code{supervised_peel()}, any 1 of the 4 components may be
plotted, e.g. \code{orig}, \code{weighted}, \code{unweighted}, or \code{peeled}.}

\item{add.apts}{Vector of additional SOMAmer reagents to highlight, beyond
the sample handling SMVs.}

\item{samples}{Character. A string corresponding to the samples to be used
to color the points in the projection plot.}

\item{filename}{Optional. File name if plot is to be saved to an
external file. Include file extension (ex. \code{filename = "example_file.png"}),
as file type will be inferred from the extension.}

\item{scale}{Optional. Scale (as a numeric value) for the saved plot.}

\item{matrix.type}{Which matrix type ("plasma" or "serum") to use for the
cell abuse SMVs. Partial matching is allowed.}

\item{dims}{Numeric. Length 2. Which dimensions of the PCA to plot. Must be
the same for the projection and rotation.}

\item{add.apt.col}{Character. The color to be used for the additional apts
specified in \code{add.apts} argument.}

\item{legend.pos}{Coordinates of the two legends, in \code{c(0, 0)} format. See Details.}
}
\value{
A sample handling plot.
}
\description{
Quick wrapper for plotting PCA sample handling, both the sample projection
and the rotation (loadings) of the protein space.
}
\details{
For the legend positions:
\tabular{rl}{
\code{c(1, 1)} \tab top right \cr
\code{c(1, 0)} \tab bottom right \cr
\code{c(0, 1)} \tab top left \cr
\code{c(0, 0)} \tab bottom left\cr
}
}
\examples{
data <- log10(sim_adat[c(1:10L, 91:100L), ])
# Create fake matching SH analytes
new_seqs <- c(
  "seq.4124.24",   # cell lysis
  "seq.2811.27",   # platelet activation
  "seq.2242.51"    # complement activation
)
names(data)[31:33L] <- new_seqs   # assign into 'noise' apts
pca <- center_scale(data, center = TRUE, scale = FALSE) |>
  strip_meta() |>
  prcomp2()

plotSampleHandling(pca, samples = data$class_response, matrix.type = "p")
plotSampleHandling(pca, samples = data$class_response, matrix.type = "p",
                   add.apts = "seq.2802.68")

# Save to file
f_out <- tempfile("SHplot-", fileext = ".pdf")
plotSampleHandling(pca, samples = data$class_response, matrix.type = "p",
                   legend.pos = c(0, 0), filename = f_out)
}
\seealso{
\code{\link[=points]{points()}}

Other PCA plots: 
\code{\link{plotPCAdims}()},
\code{\link{plotProjection}()},
\code{\link{plotRotation}()},
\code{\link{screeplotAUC}()}
}
\author{
Stu Field, Amanda Hiser
}
\concept{PCA plots}
