% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prcomp2.R
\name{prcomp2}
\alias{prcomp2}
\title{Principal Component Analysis}
\usage{
prcomp2(x, tol = NULL)
}
\arguments{
\item{x}{a numeric or complex matrix (or data frame) which provides
    the data for the principal components analysis.}

\item{tol}{a value indicating the magnitude below which components
    should be omitted. (Components are omitted if their
    standard deviations are less than or equal to \code{tol} times the
    standard deviation of the first component.)  With the default null
    setting, no components are omitted (unless \code{rank.} is specified
    less than \code{min(dim(x))}.).  Other settings for \code{tol} could be
    \code{tol = 0} or \code{tol = sqrt(.Machine$double.eps)}, which
    would omit essentially constant components.}
}
\value{
A \code{prcomp} class object.
}
\description{
Performs a principal components analysis on the given data matrix
and returns the results as a \code{prcomp} class object. This
function is "borrowed" from \code{\link[stats:prcomp]{stats::prcomp()}}, except
with some extra returned components of the \code{SVD} that are not
returned by the standard function as well as hard coding
the projection to be returned and some simple data checks.
}
\note{
We've simply hijacked \code{\link[=prcomp]{prcomp()}} and added elements to the return value.
}
\seealso{
\code{\link[=prcomp]{prcomp()}}, \code{\link[=svd]{svd()}}
}
\author{
Stu Field, Mike R. Mehan
}
