% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplot-auc.R
\name{screeplot_auc}
\alias{screeplot_auc}
\title{Plot AUC Dependent Screeplot}
\usage{
screeplot_auc(
  data.prcomp,
  auc.classes,
  auc.proj = NULL,
  main = NULL,
  nPCs = 20L,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object
returned by \code{\link[=prcomp2]{prcomp2()}}.}

\item{auc.classes}{A vector of the training classes used in calculating the AUC.}

\item{auc.proj}{An alternative projection matrix from another decomposition.
By default, the projection from \code{data.prcomp} is used.}

\item{main}{\code{character(1)}. A string containing the title for the plot.}

\item{nPCs}{\code{integer(1)}. Number of PCs to plot.}

\item{...}{Additional arguments passed to \code{\link[=geom_col]{geom_col()}}.}
}
\description{
Plot a screeplot with the variance bars colored according
to the AUC of the corresponding classes in a specified
training response vector.
}
\examples{
pca <- center_scale(pcapkg:::log10_ft(simdata), center = TRUE, scale = FALSE) |>
  feature_matrix() |>
  prcomp2()
screeplot_auc(pca, simdata$class_response, main = "My ScreePlot by AUC")
}
\seealso{
\code{\link[=screeplot]{screeplot()}}, \code{\link[=plot_scree]{plot_scree()}}

Other PCA plots: 
\code{\link{plot_pca_dims}()},
\code{\link{plot_projection}()},
\code{\link{plot_rotation}()}
}
\author{
Stu Field, Michael R. Mehan
}
\concept{PCA plots}
