% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/screeplotAUC.R
\name{screeplotAUC}
\alias{screeplotAUC}
\title{Plot AUC Dependent Screeplot}
\usage{
screeplotAUC(
  data.prcomp,
  auc.classes,
  auc.proj = NULL,
  main = NULL,
  nPCs = 20,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object returned by
\code{\link[=prcomp2]{prcomp2()}}.}

\item{auc.classes}{A vector of the training classes used in calculating the AUC.}

\item{auc.proj}{An alternative projection matrix from another decomposition.
By default, the projection from \code{data.prcomp} is used.}

\item{main}{Character. A string containing the title for the plot.}

\item{nPCs}{Integer. Number of PCs to plot. Defaults to 20.}

\item{...}{Additional arguments passed to \code{\link[=geom_col]{geom_col()}}.}
}
\description{
Plot a screeplot with the variance bars colored according
to the AUC of the corresponding classes in a specified
training response vector.
}
\examples{
pca <- center_scale(log10(sim_test_data), center = TRUE, scale = FALSE) |>
  strip_meta() |>
  prcomp2()
screeplotAUC(pca, sim_test_data$class_response, main = "My ScreePlot by AUC")
}
\seealso{
\code{\link[=screeplot]{screeplot()}}, \code{\link[=plotScree]{plotScree()}}

Other PCA plots: 
\code{\link{plotPCAdims}()},
\code{\link{plotProjection}()},
\code{\link{plotRotation}()},
\code{\link{plotSampleHandling}()}
}
\author{
Michael Mehan, Amanda Hiser
}
\concept{PCA plots}
