% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-pca-names.R
\name{get_pca_names}
\alias{get_pca_names}
\title{Get PCA Names}
\usage{
get_pca_names(x, type = c("rotation", "projection"), dim = 1L, value)
}
\arguments{
\item{x}{A \code{pca} class object. See \code{\link[=pca]{pca()}}.}

\item{type}{Matched character. Either \code{"rotation"} or \code{"projection"}.}

\item{dim}{Integer. Which PCA dimension to select.}

\item{value}{Numeric. The value corresponding to the
dimension to interrogate for names.}
}
\value{
A character string corresponding to the feature names
that exceed \code{value} in the dimension \code{dim}.
}
\description{
Determine the names of the features (rotation) or samples
(projection) above/below a specified cutoff (value) in the
principal component dimension specified by \code{dim}.
}
\examples{
feat <- grep("^seq", names(sim_adat), value = TRUE)
for (i in feat) sim_adat[[i]] <- log10(sim_adat[[i]])
pca <- pca(sim_adat)

get_pca_names(pca, "r", 1, 0.1)      # feature names

get_pca_names(pca, "p", value = 0.1) # sample names
}
\seealso{
\code{\link[=pca]{pca()}}
}
\author{
Stu Field
}
