% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-pca-dims.R
\name{plot_pca_dims}
\alias{plot_pca_dims}
\title{Plot A PCA Dimension}
\usage{
plot_pca_dims(
  data.prcomp,
  value = c(NA, "rotation", "x"),
  dims,
  classes,
  main = sprintf("PCA Plot (\%s)", ifelse(value == "x", "projection", value)),
  scores = NULL,
  col = NULL,
  xlab = NULL,
  ylab = NULL,
  legend_pos = "right",
  xlim = NULL,
  ylim = NULL,
  pt_cex = 2.5,
  pt_pch = 19,
  add_ellipse = FALSE,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object
returned by \code{\link[=prcomp2]{prcomp2()}}.}

\item{value}{Character. The type of plot to be generated. For projection
enter "x", for rotation enter "rotation".}

\item{dims}{\code{integer(2)}. Which dimensions to plot.}

\item{classes}{Optional. A vector indicating the classes of
observations (either features or samples) used for coloring the
points, depending on whether a rotation or a projection is to be
plotted. Must be the same length as the plot type, i.e. number
of samples for projection and features for rotation.}

\item{main}{\code{character(1)}. A string containing the title for the plot.}

\item{scores}{Optional statistical scores to pass through for the
coloring of the points during plotting. If a training data
set is passed and \code{scores = NULL},
then KS-distances (scores) will be calculated under the
hood and used to determine point color. Can be either a
single value (e.g. "red") or a vector of color values the
same length as the number of observations.
This overrides the point color assigned by \code{classes} above.}

\item{col}{\code{character(1)}. The color of the points. Can be either
a single value (e.g. "red") or a vector of color values the
same length as the number of observations.
This parameter overrides the point colors determined by
both the \code{classes} and \code{scores} parameters above.}

\item{xlab}{\code{character(1)}. A string indicating the label of the x axis.}

\item{ylab}{\code{character(1)}. A string indicating the label of the y axis.}

\item{legend_pos}{\code{character(1)}. Position of the legend, if plotted.
Options include "left", "right", "bottom", or "top", or "none"
if no legend is desired. Ignored if \code{value = "r"} (rotation) and
classes are not specified.}

\item{xlim}{Numeric. Optional limits for the x-axis in the
format \code{c(0, 0)}.}

\item{ylim}{Numeric. Optional limits for the y-axis in the
format \code{c(0, 0)}.}

\item{pt_cex}{\code{numeric(1)}. Character expansion for the points.}

\item{pt_pch}{\code{numeric(1)}. Shape of the points. Accepted values are 0-25.}

\item{add_ellipse}{Logical. Should an ellipse be added to the
rotation? Ignored if projection (\code{value = "x"}).}

\item{...}{Additional arguments passed to \code{\link[=geom_point]{geom_point()}}.}
}
\value{
A points plot of either the rotation or projection space.
}
\description{
Plots either the projection or the loadings (rotation) of
\code{"prcomp"} object. This is a base function that is
wrapped into driver parent functions in the plotting of
principal components, most notably \code{\link[=plot_rotation]{plot_rotation()}}
and \code{\link[=plot_projection]{plot_projection()}}.
}
\examples{
# Prepare data
pca <- center_scale(pcapkg:::log_rfu(simdata), center = TRUE, scale = FALSE) |>
  feature_matrix() |>
  prcomp2()
scores <- withr::with_seed(101, rnorm(40))

# Rotation space, with scores to define point colors
plot_pca_dims(pca, value = "r", dims = 1:2L, classes = NULL,
              scores = rnorm(40))

# Projection space, with classes to define point colors
plot_pca_dims(pca, value = "x", dims = 1:2L,
              classes = simdata$class_response)
}
\seealso{
Other PCA plots: 
\code{\link{plot_projection}()},
\code{\link{plot_rotation}()},
\code{\link{screeplot_auc}()}
}
\author{
Michael Mehan, Amanda Hiser
}
\concept{PCA plots}
