% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-rotation.R
\name{plot_rotation}
\alias{plot_rotation}
\title{Plot PCA Rotation}
\usage{
plot_rotation(
  data.prcomp,
  dims = 1:2L,
  classes = NULL,
  scores = NULL,
  col = NULL,
  aptamers = NULL,
  aptamers2 = NULL,
  aptamers3 = NULL,
  aptamers4 = NULL,
  aptamers5 = NULL,
  lab.cex = 3,
  pt.cex = 2.5,
  auto.ident = TRUE,
  ...
)
}
\arguments{
\item{data.prcomp}{A prcomp class object. Typically the object returned by
\code{\link[=prcomp2]{prcomp2()}}.}

\item{dims}{Integer. Length 2. Which dimensions to plot.}

\item{classes}{Optional. A vector indicating the classes of features used for
coloring the points. Must be the same length as the number of features.}

\item{scores}{Optional. KS scores to pass through for the coloring
of the points during plotting. If a training data set is passed and
\code{scores = NULL}, then KS-distances (scores) will be calculated under the
hood and used to determine point color. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This overrides the point color assigned by \code{classes} above.}

\item{col}{Character. The color of the points. Can be either a single value
(e.g. "red") or a vector of color values the same length as the number of
observations. This parameter overrides the point colors determined by
both the \code{classes} and \code{scores} parameters above.}

\item{aptamers}{Optional. A vector of \code{SeqIds} to mark on the plot. Typically
a vector of analytes related to sample handling, or of other clinical
relevance. Marked with a hollow diamond (see \code{\link[=pch]{pch()}}).}

\item{aptamers2}{Optional. An additional vector of \code{SeqIds} to mark
on the plot, separately from those in \code{aptamers}. Marked with a hollow
triangle (see \code{\link[=pch]{pch()}}).}

\item{aptamers3}{Optional. An additional vector of \code{SeqIds} to mark
on the plot, separately from those in \code{aptamers}. Marked with a hollow
square (see \code{\link[=pch]{pch()}}).}

\item{aptamers4}{Optional. An additional vector of \code{SeqIds} to mark
on the plot, separately from those in \code{aptamers}. Marked with a hollow
circle (see \code{\link[=pch]{pch()}}).}

\item{aptamers5}{Optional. An additional vector of \code{SeqIds} to mark
on the plot, separately from those in \code{aptamers}. Marked with a hollow
upside-down triangle (see \code{\link[=pch]{pch()}}).}

\item{lab.cex}{Numeric. Font size of the point labels generated by \code{auto.ident}.
\code{auto.ident} must be TRUE.}

\item{pt.cex}{Numeric. Character expansion for the points.}

\item{auto.ident}{Logical. Should the top 10 features, in both dimensions,
be identified and labeled on the plot?}

\item{...}{Additional arguments passed to \code{\link[=plot_pca_dims]{plot_pca_dims()}}.}
}
\description{
Plot the features from a PCA rotation in a 2-dimensional scatter plot.
\strong{Note:} The precedence for the ordering of the coloring is as follows:
\enumerate{
\item{\code{col}}
\item{\code{scores}}
}
}
\examples{
feat <- grep("^seq", names(sim_adat), value = TRUE)
for (i in feat) sim_adat[[i]] <- log10(sim_adat[[i]])
pca <- center_scale(sim_adat, center = TRUE, scale = FALSE) |>
  strip_meta() |>
  prcomp2()
plot_rotation(pca, col = "green")

class <- withr::with_seed(123, sample(sim_adat$class_response, 40L))
plot_rotation(pca, classes = class)

apts <- withr::with_seed(123, sample(pcapkg:::get_analytes(sim_adat), 10L))
plot_rotation(pca, aptamers = apts)
}
\seealso{
Other PCA plots: 
\code{\link{plot_pca_dims}()},
\code{\link{plot_projection}()},
\code{\link{plot_sample_handling}()},
\code{\link{screeplot_auc}()}
}
\author{
Michael Mehan
}
\concept{PCA plots}
